refactor(transactions): split webhook by concern + zero-alloc id parse
Follow-up on the dispatch refactor: address the file-level SRP gap and the
parsing allocations.

- Split the 613-line server.go by responsibility: server.go keeps the HTTP
  controller + orchestration, tebexevent.go owns the Tebex wire shapes +
  classification + parsing, signature.go owns the HMAC verification.
- rawUint no longer allocates a json.Decoder + bytes.Reader per call; it
  trims/unquotes in place and parses with strconv (0 allocs/op, benchmarked,
  down from several).
- Fold failEvent's duplicated save-error response into saveError.

Behavior unchanged: all 15 webhook tests pass, gofmt/vet clean.
